+++ demo/src/sandboxes/leva-transient/src/App.jsx
@@ -1,6 +1,6 @@
 import React, { useRef } from 'react'
 import { useControls } from 'leva'
-import { Canvas } from 'react-three-fiber'
+import { Canvas } from '@react-three/fiber'
 import { OrbitControls } from '@react-three/drei'
 
 import * as THREE from 'three'