+++ packages/leva/CHANGELOG.md
@@ -1,5 +1,26 @@
 # leva
 
+## 0.9.12
+
+### Patch Changes
+
+- d5a5dd2: - Fix: prevent selection on Firefox when closing the Color picker.
+  - Fix: add suffix to Number plugin initial value.
+  - Feat: add `title` attribute to Labels when there is no `hint` prop.
+
+## 0.9.11
+
+### Patch Changes
+
+- cedd742: chore: update deps
+
+## 0.9.10
+
+### Patch Changes
+
+- 16e3c14: feat: add clamp Range and Interval initial values to bounds.
+  feat: when title is empty show the drag dots.
+
 ## 0.9.9
 
 ### Patch Changes