+++ packages/plugin-plot/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @leva-ui/plugin-plot
 
+## 0.9.12
+
+### Patch Changes
+
+- d5a5dd2: - Fix: prevent selection on Firefox when closing the Color picker.
+  - Fix: add suffix to Number plugin initial value.
+  - Feat: add `title` attribute to Labels when there is no `hint` prop.
+- Updated dependencies [d5a5dd2]
+  - leva@0.9.12
+
+## 0.9.10
+
+### Patch Changes
+
+- 3570975: chore: update core dependency.
+- Updated dependencies [16e3c14]
+  - leva@0.9.10
+
 ## 0.9.8
 
 ### Patch Changes